Backend/patent_search/requirements.txt
attrs>=19.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.9.3
certifi==2021.10.8
chardet==4.0.0
ConfigArgParse==1.5.3
configparser==4.0.2
elasticsearch>=7.10.1
entrypoints==0.3
idna==2.10
keyring==18.0.1
numpy==1.24.2
pandas==1.5.3
pyasn1==0.4.8
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.27.1
six==1.16.0
soupsieve==1.9.6
urllib3==1.26.13
Flask==2.1.1
sentence-transformers
faiss-cpu
Flask-Paginate
google-generativeai